requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf_yasg==1.21.8
django-jet-4
django-cors-headers